requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0 